Added multiple configuration files and updated Python notes file
This commit includes the addition of various IntelliJ-based XML files which define important configuration settings for the Python development project. The configurations manage aspects like the database, Python SDK settings, the version control system, as well as inspection profiles, to name a few.

The diff also includes the addition of a new '.gitignore' file. This was done to make sure that certain files, which are not necessary for version control, are ignored by Git.

Additionally, we've also made certain updates to the '12 Python (OOP) - dziedziczenie.ipynb' file. These changes mostly involve adding execution counts to the cells which previously did not have any. This will help in keeping track of the order of execution of the different cells in the Jupyter Notebook.

This commit is intended to streamline the development process by setting up the necessary development configurations and keeping the repository clean, as well as improving the readability of the notebook file.