Update packages in Image Processing Software notebook
Upgraded 'numpy', 'opencv-python' and 'matplotlib' packages in Image Processing Software notebook to ensure the use of latest package versions. Execution statements were also included to demonstrate the change in package versions and to show the successful installation of each package. The 'cv2' library was also tested after the packages were updated. This improves the notebook's functionality by ensuring compatibility with the latest packages and libraries.