Update execution counts and outputs in 001_ds_NumPy.ipynb
Execution counts in the Jupyter notebook '001_ds_NumPy.ipynb' were updated for several code cells to reflect the correct sequence of code execution. This creates a clear and accurate record of the execution order for future reference.
Additionally, several outputs were also updated to maintain consistency with the changes carried out on the corresponding code cells. This is crucial for correctness and replicability of the computational tasks/methods illustrated in the notebook.