Update NumPy guide with package upgrade cell
Added a new cell that upgrades NumPy package using pip before continuing with the guide. This is to make sure that the reader is using the latest version of NumPy which could help prevent any version issue. Additionally, there were updates made in the code cells throughout the notebook to improve efficiency and clarity. The execution counts have been adjusted to match the new flow of the notebook.