Update Python and package versions in notebooks
Updated Python version to 3.11.5 in `2 Image Processing Software - Tasks` notebook. Python version update was necessary to ensure compatibility with the latest libraries and tools. In `1 Image Processing Software` notebook, reduced the execution count numbers, updated numpy, opencv-python and matplotlib installation outputs. These reflect the newer versions of numpy, opencv-python and matplotlib compatible with the updated Python version. Also, to reflect that the libraries are installed system-wide and not just for the current user.