"Clean up jupyter notebook and remove output results
The outputs from various cells in the jupyter notebook 'Python (programowanie funkcyjne) - funkcje' were removed. This was done to ensure that the notebook is clean and easily readable. Any user running this notebook now will be able to run cells independently and get the output based on their environment setup. Also, an image link that was previously visible was changed to a comment."