"Revise content and simplify package dependencies in Digital Image Filters exercise"
Made two key changes in this commit.

Firstly, a minor formatting change was made to the description of a high pass filter in the Digital Image Filters exercise. This included removing an unnecessary tab for better readability.

Secondly, the list of dependencies that are printed out was greatly simplified. This reduces clutter and makes the notebook easier to understand. Removed dependencies were either unused in the exercise, or implicitly imported by other, higher-level packages. The change will also help to reduce the notebook's load time.