Add introduction HTML file for Machine Learning
This commit includes the addition of a new HTML file, '00_ML_Wprowadzenie.html', which serves as an introduction to the Machine Learning directory. The file includes extensive CSS and Bootstrap formatting for improved readability and usability. This addition is part of an effort to enhance the comprehensiveness and structure of the Machine Learning resources.