Update Python package versions and refresh data responses
The python package versions for pip and setuptools in Cloud-Based Image Processing notebook have been updated to reflect the latest versions used. The data and its responses have also been refreshed with updated values. This is due to the usage of real-time data, which could have been updated or changed since the last commit, ensuring the tutorial reflects the current situation and making it more accurate and up-to-date for users. Additionally, Logical and Arithmetical Operations on Images - Exercises notebook's execution count was set to null to not mention the specific amount of times the code has been run while developing or testing, thus providing a clean state for future users or collaborators.